backend/requirements.txt
# FastAPI 服务
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# 爬虫引擎
httpx==0.27.0
playwright==1.47.0
beautifulsoup4==4.12.0
lxml==5.3.0

# 验证码识别
ddddocr==1.5.0
Pillow==10.4.0

# 数据库
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
redis[hiredis]==5.1.0
alembic==1.13.0

# AI / LLM
openai==1.50.0

# SSE 流式响应
sse-starlette==2.1.0

# 工具
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
loguru==0.7.2
